cosmetic changes
1.
Summary: Class 'X' has a constructor with 1 argument that is not explicit.

Message: Class 'X' has a constructor with 1 argument that is not explicit. Such constructors should in general be explicit for type safety reasons. Using the explicit keyword in the constructor means some mistakes when using the class can be avoided.

more info: https://en.cppreference.com/w/cpp/language/explicit

2.
Summary: The function 'raise' overrides a function in a base class but is not marked with a 'override' specifier.
Message: The function 'raise' overrides a function in a base class but is not marked with a 'override' specifier.

more info: https://en.cppreference.com/w/cpp/language/override