Improve long double serialization
The "long double" type is not very portable, as it can have different
sizes on different architectures. While there is some ad-hoc fix for
Windows and ARM, the code is not really portable nor works
cross-architecturally.

This patch adds a check for the actual size of long double and tries to
convert to the __float128 type for serialization if long double is
shorter than 16 bytes. If __float128 is not available and long double
has 8 bytes, the previous fallback is implemented, which puts a regular
zero-padded double into the field.

Unlike the old code, the new serialization logic works on x86 and
certain PowerPC architectures, where long double has 12 bytes. The code
could be improved further by a proper, manually coded conversion between
double precision and quadruple precision if __float128 is not available
(the "Windows" case).

Signed-off-by: Timo Röhling <timo@gaussglocke.de>