Prevent unexpected length_error exception
On some architectures, std::vector<bool> has a max_size() below 2^32-1,
so the early resize() call will throw a length_error exception before
the deserialization code has a chance to throw the
NotEnoughMemoryException. This patch moves the resize() operation after
the length check, so the correct exception is thrown consistently.

Signed-off-by: Timo Röhling <timo@gaussglocke.de>