Fix a compile warning with g++ 13.2.0.
When building the tests with g++ 13.2.0, it complains
that the EXPECT_LONG_DOUBLE_EQ macro has no effect.
That ends up being true; because it is just a x == y,
with no EXPECT statement around it, it actually does
nothing.

Here, I just remove this macro because it is unnecessary;
we can just do EXPECT_TRUE(x == y) and get the same effect
with less code.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>